% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bimodality.R
\name{apcProts_function}
\alias{apcProts_function}
\title{Convert protein data txt file into a dataframe}
\usage{
apcProts_function(txt_file)
}
\arguments{
\item{txt_file}{A txt file.}
}
\value{
A data frame.
}
\description{
Convert protein data txt file into a dataframe
}
\examples{
ApcProts_raw <- apcProts_function(Apc_prots_txt)
}
