% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bimodality.R
\name{assayData_function}
\alias{assayData_function}
\title{Convert assay data txt file to data frame}
\usage{
assayData_function(txt_file)
}
\arguments{
\item{txt_file}{A txt file.}
}
\value{
A data frame.
}
\description{
Convert assay data txt file to data frame
}
\examples{
assayData(txt_file)
}
