% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bimodality.R
\name{CalculateTorque}
\alias{CalculateTorque}
\title{Calculate the torque of the difference in CDFs.}
\usage{
CalculateTorque(rho_t, ApcLocs, pow)
}
\arguments{
\item{rho_t}{correlation matrix for a network to all genes.}

\item{ApcLocs}{position of the APC targets.}
}
\value{
torque value.
}
\description{
Calculate the torque of the difference in CDFs.
}
\examples{
D[m] = CalculateTorque(rho.t[[p]], r_group, 0)
}
