% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bimodality.R
\name{collapseProbes}
\alias{collapseProbes}
\title{Reduce the eSet variable into eSet collapsed variable with either mean or variance method}
\usage{
collapseProbes(
  x,
  method = "mean",
  topTable = NULL,
  colNameOfStat = NULL,
  colNameOfGeneSymbol = "SYMBOL"
)
}
\arguments{
\item{x}{An eSet variable.}

\item{method}{mean, or variance.}

\item{colNameOfGeneSymbol}{name of the column used to reduce the eSet.}
}
\value{
eSet variable.
}
\description{
Reduce the eSet variable into eSet collapsed variable with either mean or variance method
}
\examples{
eset.collapsed <- eset.collapsed_function(eset.raw)
}
