% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bimodality.R
\name{eset.collapsed_function}
\alias{eset.collapsed_function}
\title{Convert eSet data by normalization of assay parameter using collapseProbes function}
\usage{
eset.collapsed_function(eset.raw)
}
\arguments{
\item{eset.raw}{An eSet variable.}
}
\value{
Another eSet variable, that has normalized its assay data parameter.
}
\description{
Convert eSet data by normalization of assay parameter using collapseProbes function
}
\examples{
eset.collapsed_function <- function(eset.raw)
}
