% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bimodality.R
\name{eset.raw_function}
\alias{eset.raw_function}
\title{Combines assay, feature, pheno data into a single variable the type of eSet}
\usage{
eset.raw_function(Apc_dChip_matrix, Apc_dChip, p_data, Apc_p21_ids)
}
\arguments{
\item{Apc_dChip_matrix}{assay data stored in matrix type.}

\item{Apc_dChip}{assay data stored in data frame type.}

\item{p_data}{pheno data stored in data frame type.}

\item{Apc_p21_ids}{feature data stored in data frame type.}
}
\value{
A eSet variable.
}
\description{
Combines assay, feature, pheno data into a single variable the type of eSet
}
\examples{
eset.raw_function <- function(Apc_dChip_matrix, Apc_dChip, p_data, Apc_p21_ids)
}
