% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bimodality.R
\name{featureData_function}
\alias{featureData_function}
\title{Convert feature data txt file to data frame}
\usage{
featureData_function(txt_file)
}
\arguments{
\item{txt_file}{A txt file.}
}
\value{
A data frame.
}
\description{
Convert feature data txt file to data frame
}
\examples{
featureData(txt_file)
}
