% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bimodality.R
\name{main_DGEList_function}
\alias{main_DGEList_function}
\title{calculates the p_value, and bimodality score by utilizing the functions provided above}
\usage{
main_DGEList_function(
  DGEList,
  Apc_prots_txt,
  petal_networks_txt,
  maxperm_value
)
}
\arguments{
\item{DGEList}{DGEList variable containing the mRNA target.}

\item{Apc_prots_txt}{protein networks stored.}

\item{petal_networks_txt}{petal networks stored in a txt file.}

\item{maxperm_value}{integer to be provided for permutations of the calculation.}
}
\value{
a data frame where p values and bimodality scores are stored in.
}
\description{
calculates the p_value, and bimodality score by utilizing the functions provided above
}
\examples{
bimodality_DGE_df <- main_DGEList_function(sample_DGElist, "Apc prots.txt", "petal_network.txt", 1000)
}
