% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bimodality.R
\name{main_function}
\alias{main_function}
\title{calculates the p_value, and bimodality score by utilizing the functions provided above}
\usage{
main_function(
  APC_RMA_variable,
  Apc_p21_variable,
  pdata_variable,
  Apc_prots_txt,
  petal_networks_txt,
  maxperm_value
)
}
\arguments{
\item{APC_RMA_variable}{assay data in a data frame class.}

\item{Apc_p21_variable}{feature data in a data frame class.}

\item{pdata_variable}{pheno data in a data frame class.}

\item{Apc_prots_txt}{protein networks stored.}

\item{petal_networks_txt}{petal networks stored in a txt file.}

\item{maxperm_value}{integer to be provided for permutations of the calculation.}
}
\value{
a data frame where p values and bimodality scores are stored in.
}
\description{
calculates the p_value, and bimodality score by utilizing the functions provided above
}
\examples{
bimodality_df <- main_function("Apc dChip.txt", "Apc p21 - ids.txt", "pdata.txt", "Apc prots.txt", "petal_network.txt", 1000)
}
