% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bimodality.R
\name{mapGmt2NamedList}
\alias{mapGmt2NamedList}
\title{Turn a txt file into a GMT one.}
\usage{
mapGmt2NamedList(gmtFileName = NULL, type = "simple")
}
\arguments{
\item{gmtFileName}{A txt file.}
}
\value{
a gmt file.
}
\description{
Turn a txt file into a GMT one.
}
\examples{
petal_networks_list <- mapGmt2NamedList(petal_networks_txt, type = "broad")
}
