% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bimodality.R
\name{mRNA_DGEList_function}
\alias{mRNA_DGEList_function}
\title{Assigns the provided DGEList variable into mRNA variable}
\usage{
mRNA_DGEList_function(sample_DGElist)
}
\arguments{
\item{sample_DGElist}{A DGEList variable.}
}
\value{
A mRNA variable.
}
\description{
Assigns the provided DGEList variable into mRNA variable
}
\examples{
mRNA <- mRNA_DGEList_function(DGEList)
}
