% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bimodality.R
\name{mRNA_function}
\alias{mRNA_function}
\title{Create a mrna variable by merging assay data of eSet and feature data}
\usage{
mRNA_function(eset.collapsed, Apc_p21_ids)
}
\arguments{
\item{eset.collapsed}{An eSet variable.}

\item{Apc_p21_ids}{feature data.}
}
\value{
mrna variable.
}
\description{
Create a mrna variable by merging assay data of eSet and feature data
}
\examples{
mRNA_function <- function(eset.collapsed, Apc_p21_ids)
}
