% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bimodality.R
\name{phenoData_function}
\alias{phenoData_function}
\title{Convert pheno data txt file to data frame}
\usage{
phenoData_function(txt_file)
}
\arguments{
\item{txt_file}{A txt file.}
}
\value{
A data frame.
}
\description{
Convert pheno data txt file to data frame
}
\examples{
phenoData(txt_file)
}
