% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bimodality.R
\name{tscore_function}
\alias{tscore_function}
\title{Calculate the tScore of the provided mRNA variable}
\usage{
tscore_function(mRNA)
}
\arguments{
\item{mRNA}{A mRNA variable.}
}
\value{
tScore of the provided argument.
}
\description{
Calculate the tScore of the provided mRNA variable
}
\examples{
tscore <- tscore_function(mRNA)
}
